api/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
django-filter==23.5
folium==0.19.2
